requirements.txt
boto3==1.34.51
yt-dlp==2024.3.10
python-dotenv==1.0.1